Fix: GridInfo properties
- Delete old grid files, avoid confused again
