+++ module/map_detection/grid_info.py
@@ -52,6 +52,8 @@ class GridInfo:
     is_carrier = False
     is_movable = False
     cost = 9999
+    cost_1 = 9999
+    cost_2 = 9999
     connection = None
     weight = 1
 
@@ -132,6 +134,14 @@ class GridInfo:
     @property
     def is_accessible(self):
         return self.cost < 9999
+
+    @property
+    def is_accessible_1(self):
+        return self.cost_1 < 9999
+
+    @property
+    def is_accessible_2(self):
+        return self.cost_2 < 9999
 
     @property
     def is_nearby(self):
